pyproject.toml
[tool.poetry]
name = "fastapi-skeleton"
version = "0.1.0"
description = ""
authors = ["lifangjun"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.115.0"
uvicorn = "0.31.0"
loguru = "0.7.2"
pydantic = ">=2.9.2,<2.10.0"
pydantic-settings = "2.5.2"
poetry = "1.8.3"
poetry-add-requirements-txt = "1.4.0"
sqlalchemy = "2.0.35"
pymysql = "1.1.1"
mysqlclient = "2.2.4"
cryptography = "43.0.1"
pyjwt = "2.9.0"
passlib = "1.7.4"
bcrypt = "4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
